Version 1.6.0
Updated modeling of the source term-it is now modeled as a "circular wave" in x and y with harmonics at 2.4 GHz and 5.0 GHz. The problem is then reformulated as solving the vector-valued Helmholtz equation for the x- and y- components of the electric field in the phasor domain.  The density of the FEM is increased, the FEM system now solved is approximately of dimension 10,200 x 10,200, and is solved using SciPy's sparse matrix subroutines.